Added support for elasticsearch 2.x.x
Added new pom.xml.elasticsearch-2.0.0 example for building with
elasticsearch 2.0.0 (tested on 2.0.0 and 2.0.1 so far). And added a new
Elasticsearch2Emitter.java that contains all the needed changes to make
this library work with elasticsearch 2.x.x.

//TODO - The changes in Elasticsearch2Emitter.java need to be merged
into ElasticsearchEmitter.java and the imort of "import
org.elasticsearch.common.settings.ImmutableSettings;" needs to be driven
by the version of elasticsearch used, the "ElasticsearchEmitter"
function also needs to be dependent on the version.